Add crossorigin attribute to remote CSS imports
https://www.npmjs.com/package/ember-cli-sri#crossorigin-attribute

When the request doesn't match Same Origin Policy the crossorigin
attribute MUST be present for the integrity of the file to be checked.
With an integrity set on an external origin and a missing crossorigin
the browser will choose to 'fail-open' which means it will load the
resource as if the integrity attribute was not set.

Signed-off-by: Anton Arapov <arapov@gmail.com>